+++ jklib/django/db/validators.py
@@ -3,38 +3,45 @@
 
 # Django
 from django.core.exceptions import ValidationError
+from django.utils.deconstruct import deconstructible
 
 
 # --------------------------------------------------------------------------------
-# > Validator
+# > Validators
 # --------------------------------------------------------------------------------
-def length_validator(min_=None, max_=None, trim=True):
-    """
-    Checks that the field length is within the provided limits
-    :param int min_: The min length
-    :param int max_: The max length
-    :param bool trim: Whether to trim the value before the check. Defaults to True.
-    :return: The check function
-    :rtype: function
-    """
+@deconstructible
+class LengthValidator:
+    """Validator that can check the min and/or max length of a field"""
 
-    def check(value):
+    message = ""  # Depends on args
+
+    def __init__(self, min_=None, max_=None, trim=True):
         """
-        Performs the actual check with the provided arguments
-        :raises ValidationError: If the text is not within the allowed ranges
+        Sets the instance attributes
+        :param int min_: The min length
+        :param int max_: The max length
+        :param bool trim: Whether to trim the value before the check. Defaults to True.
         """
-        if trim:
+        self.min_ = min_
+        self.max_ = max_
+        self.trim = trim
+
+    def __call__(self, value):
+        """
+        Checks if the value meets the length criterion
+        :param str value: The current string value of the field
+        :raises ValidationError: If the field is either too short or too long
+        """
+        if self.trim:
             value = value.strip()
         length = len(value)
         # Check min length
-        if min_ is not None:
-            if length < min_:
-                raise ValidationError(
-                    f"The text must be at least {min_}-character long"
-                )
+        if self.min_ is not None:
+            if length < self.min_:
+                self.message = f"The text must be at least {self.min_}-character long"
+                raise ValidationError(self.message)
         # Check max length
-        if max_ is not None:
-            if length > max_:
-                raise ValidationError(f"The text must be at most {max_}-character long")
-
-    return check
+        if self.max_ is not None:
+            if length > self.max_:
+                self.message = f"The text must be at most {self.max_}-character long"
+                raise ValidationError(self.message)